Split the ledger store into rows/queries/store; drop dead barrel
store.ts had grown to 446 lines mixing three concerns. Split into three
focused files, each well under 200 lines:

- rows.ts — the one place that knows the table columns and row mapping,
  plus the small readers both other files build on.
- queries.ts — read API: transfers-by-account/event, and the
  SQL-aggregate balance reads.
- store.ts — the write path only (idempotent post + replay/conflict and
  reversal guards).

Also delete src/shared/ledger/mod.ts, an unused re-export barrel (no
importers), and reword the denser comments in plainer language.

100% coverage, 0% duplication, lint and typecheck clean.
