+++ test/lib/code-quality.test.ts
@@ -512,6 +512,7 @@ describe("code quality", () => {
       "shared/ledger/reverse.ts",
       "shared/ledger/reconcile.ts",
       "shared/accounting/store.ts",
+      "shared/accounting/queries.ts",
       "shared/accounting/mappers.ts",
     ];
